fix(providers): relocate NotificationProvider
